refactoring docs, turning both platforms.py files back into a single file
