+++ docs/FAQ.rst
@@ -19,7 +19,7 @@ Common Errors
 .. dropdown:: **"Manager only - must be at least one worker (2 MPI tasks)" when running with multiprocessing and multiple workers specified.**
 
   If your code was recently switched from MPI to multiprocessing,
-  make sure that :class:`libE_specs<libensemble.specsV2.LibeSpecs>` is populated
+  make sure that :class:`libE_specs<libensemble.specs.LibeSpecs>` is populated
   with ``"comms": "local"`` and ``"nworkers": [int]``.
 
 .. dropdown:: **"AssertionError: alloc_f did not return any work, although all workers are idle."**
