+++ docs/data_structures/gen_specs.rst
@@ -32,7 +32,7 @@ Can be constructed and passed to libEnsemble as a Python class or a dictionary.
         )
         ...
 
-    .. autopydantic_model:: libensemble.specsV2.GenSpecs
+    .. autopydantic_model:: libensemble.specs.GenSpecs
       :model-show-json: False
       :model-show-config-member: False
       :model-show-config-summary: False
