+++ docs/tutorials/local_sine_tutorial.rst
@@ -62,7 +62,7 @@ need to write a new allocation function.
               sizes are placed within the ``gen_specs["user"]`` dictionary, while input/output and other fields
               that libEnsemble needs to operate the generator are placed outside ``user``.
 
-        Later on, we'll populate :class:`gen_specs<libensemble.specsV2.GenSpecs>` and ``persis_info`` when we initialize libEnsemble.
+        Later on, we'll populate :class:`gen_specs<libensemble.specs.GenSpecs>` and ``persis_info`` when we initialize libEnsemble.
 
         For now, create a new Python file named ``generator.py``. Write the following:
 
@@ -129,7 +129,7 @@ need to write a new allocation function.
         Next, we'll write our simulator function or :ref:`sim_f<api_sim_f>`. Simulator
         functions perform calculations based on values from the generator function.
         The only new parameter here is :ref:`sim_specs<datastruct-sim-specs>`, which
-        serves a purpose similar to the :class:`gen_specs<libensemble.specsV2.GenSpecs>` dictionary.
+        serves a purpose similar to the :class:`gen_specs<libensemble.specs.GenSpecs>` dictionary.
 
         Create a new Python file named ``simulator.py``. Write the following:
