+++ libensemble/specs.py
@@ -7,9 +7,6 @@ from pydantic import BaseModel, Field
 from libensemble.alloc_funcs.give_sim_work_first import give_sim_work_first
 from libensemble.resources.platforms import Platform
 
-_UNRECOGNIZED_ERR = "Unrecognized field. Check closely for typos, or libEnsemble's docs"
-_OUT_DTYPE_ERR = "Unable to coerce '{}' into a NumPy dtype. It should be a list of 2-tuples or 3-tuples"
-
 __all__ = ["SimSpecs", "GenSpecs", "AllocSpecs", "ExitCriteria", "LibeSpecs", "_EnsembleSpecs"]
 
 
@@ -141,16 +138,16 @@ class ExitCriteria(BaseModel):
     Specifications for configuring when libEnsemble should stop a given run.
     """
 
-    sim_max: Optional[int]
+    sim_max: Optional[int] = None
     """Stop when this many new points have been evaluated by simulation functions."""
 
-    gen_max: Optional[int]
+    gen_max: Optional[int] = None
     """Stop when this many new points have been generated by generator functions."""
 
-    wallclock_max: Optional[float]
+    wallclock_max: Optional[float] = None
     """Stop when this many seconds has elapsed since the manager initialized."""
 
-    stop_val: Optional[Tuple[str, float]]
+    stop_val: Optional[Tuple[str, float]] = None
     """Stop when ``H[str] < float`` for the given ``(str, float)`` pair."""
 
 
@@ -162,7 +159,7 @@ class LibeSpecs(BaseModel):
     comms: Optional[str] = "mpi"
     """ Manager/Worker communications mode. ``'mpi'``, ``'local'``, ``'local_threading'``, or ``'tcp'`` """
 
-    nworkers: Optional[int]
+    nworkers: Optional[int] = 0
     """ Number of worker processes in ``"local"`` or ``"tcp"``."""
 
     mpi_comm: Optional[Any] = None
@@ -357,7 +354,7 @@ class LibeSpecs(BaseModel):
     stats_fmt: Optional[dict] = {}
     """ Options for formatting ``'libE_stats.txt'``. See 'Formatting libE_stats.txt'. """
 
-    workers: Optional[List[str]]
+    workers: Optional[List[str]] = []
     """ TCP Only: A list of worker hostnames. """
 
     ip: Optional[str] = None
@@ -369,10 +366,10 @@ class LibeSpecs(BaseModel):
     authkey: Optional[str] = f"libE_auth_{random.randrange(99999)}"
     """ TCP Only: Authkey for Manager's system."""
 
-    workerID: Optional[int]
+    workerID: Optional[int] = 0
     """ TCP Only: Worker ID number assigned to the new process. """
 
-    worker_cmd: Optional[List[str]]
+    worker_cmd: Optional[List[str]] = []
     """
     TCP Only: Split string corresponding to worker/client Python process invocation. Contains
     a local Python path, calling script, and manager/server format-fields for ``manager_ip``,
@@ -397,20 +394,20 @@ class LibeSpecs(BaseModel):
     and/or assignment of resources to workers. ``"resource_info"`` will be ignored.
     """
 
-    num_resource_sets: Optional[int]
+    num_resource_sets: Optional[int] = 0
     """
     Total number of resource sets. Resources will be divided into this number.
     If not set, resources will be divided evenly (excluding zero_resource_workers).
     """
 
-    gen_num_procs: Optional[int]
+    gen_num_procs: Optional[int] = 0
     """
     The default number of processors (MPI ranks) required by generators. Unless
     overridden by the equivalent `persis_info` settings, generators will be
     allocated this many processors for applications launched via the MPIExecutor.
     """
 
-    gen_num_gpus: Optional[int]
+    gen_num_gpus: Optional[int] = 0
     """
     The default number of GPUs required by generators. Unless overridden by
     the equivalent `persis_info` settings, generators will be allocated this
@@ -474,5 +471,5 @@ class _EnsembleSpecs(BaseModel):
     alloc_specs: Optional[AllocSpecs]
     """ Specifications for the allocation function. """
 
-    nworkers: Optional[int]
+    nworkers: Optional[int] = 0
     """ Number of worker processes to spawn (only in local/tcp modes). """
